requirements.txt
streamlit
openai
pypdf
pillow
pytesseract
